Fix the recto/verso shuffle dance
The "ebook" option for memoir is a bit of a misnomer. You would think it sets
things up for electronic books but it actually sets up a traditional book with
the form factor of an e-reader. As a result of this the recto and verso
(odd and even) pages have different margins. I think it is a bit distracting
because the text shifts a little with each page. I don't know if this will show
up in the pull request correctly but the effect is:

       RV Shuffle            oneside
    +---------------+   +---------------+
    |    =========  |   |   =========   |
    |    =========  |   |   =========   |
    |    =========  |   |   =========   |
    |    =========  |   |   =========   |
    |    =========  |   |   =========   |
    |    =========  |   |   =========   |
    |    =========  |   |   =========   |
    |    =========  |   |   =========   |
    +---------------+   +---------------+
    +---------------+   +---------------+
    |  =========    |   |   =========   |
    |  =========    |   |   =========   |
    |  =========    |   |   =========   |
    |  =========    |   |   =========   |
    |  =========    |   |   =========   |
    |  =========    |   |   =========   |
    |  =========    |   |   =========   |
    |  =========    |   |   =========   |
    +---------------+   +---------------+
    +---------------+   +---------------+
    |    =========  |   |   =========   |
    |    =========  |   |   =========   |
    |    =========  |   |   =========   |
    |    =========  |   |   =========   |
    |    =========  |   |   =========   |
    |    =========  |   |   =========   |
    |    =========  |   |   =========   |
    |    =========  |   |   =========   |
    +---------------+   +---------------+

This is alluded to on page 43 of the memman document but the author does
not really describe what is going on.