Provide backrefs in bibliography as in glossaries
The bibliography entries now provide a page backref just like the
glossary entries. It was supposed to be as simple as passing the
option `pagebackref` to hypreref. For some reason this caused a
problem with two URLs that did not escape the `#` character. Why
this was not a problem before is beyond me. I checked and the
links work--are not escaped--in the resulting pdf. Finally because
there were not enough latex build process artifacts this added a
Crypto.brf file to the build aftermath. Added an entry to gitignore
to ignore the Crypto.brf file.

I hope all is well in the land of the lakenois.