RC4 attack: merge two paragraphs
As it stands, these two paragraphs read as though they're independent
from each other, but actually I suspect the second paragraph is intended
to continue the arc of the first paragraph. (At least the second
paragraph seems to depend on the first.)

This is the easiest fix. If the paragraph ends up too long then maybe
a connecting sentence would be better.